sensor as module
