+++ tests/fixtures.py
@@ -1,0 +1,1 @@
+import pytest
